refactor(eks): rename utils.py to a role-named aws_credentials module
The file only maps stored AWS credentials to botocore's PascalCase shape. Name it by that role and delete the utils grab-bag.
